// src/models/parqetRecord.ts
export class ParqetRecord {
    dateTime: Date;
    date: Date;
    time: string;
    price: number;
    shares: number;
    amount: number;
    tax: number;
    fee: number;
    realizedGains: number;
    type: string;
    broker: string;
    assetType: string;
    identifier: string;
    wkn: string;
    originalCurrency: string;
    currency: string;
    fxRate: number;
    holding: string;
    holdingName: string;
    holdingNickname: string;
    exchange: string;
    avgHoldingPeriod: string
}